feat: add support for storing multiple images in service records
The code changes add a new column "images" to the "service_records" table in the database. This column will store an array of image URLs associated with each service record. This enhancement allows users to attach multiple images to a service record.

Note: The suggested commit message follows the conventional format of starting with a type (feat for feature) and providing a concise description of the changes made.